Back/requirements.txt
anyio==4.10.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
Deprecated==1.2.18
exceptiongroup==1.3.0
firebase_admin==7.1.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.12
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.1
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
qrcode==8.2
requests==2.32.4
rich==13.9.4
rsa==4.9.1
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
